\name{sample_data}
\docType{data}
\alias{sample_data}
\title{Sample Data from Jacobson & Truax (1991)}
\description{
  This data set is an excerpt from Table 2 of Jacobson & Truax (1991).
}
\usage{sample_data}
\format{A CSV table containing the columns `ppid`, `pre` and `post` where `ppid`
is a continuously incrementing list of unique integers, `pre` contains pretest
values (floating-point) and `post` contains posttest values (floating-point
too)}
\source{Table 2 in Jacobson & Truax (1991)}
\references{
Jacobson, N. S., & Truax, P. (1991).
Clinical Significance: A Statistical Approach to Defining Meaningful Change in Psychotherapy Research.
Journal of Consulting and Clinical Psychology, 59, 12-19.
<doi:10.1037/0022-006X.59.1.12>
}
\keyword{datasets}
